+++ lib/WebpackOptionsApply.js
@@ -59,11 +59,8 @@ const URLPlugin = require("./dependencies/URLPlugin");
 
 const WorkerPlugin = require("./dependencies/WorkerPlugin");
 
-const ConstValueParserPlugin = require("./javascript/ConstValueParserPlugin");
-
 const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 const JavascriptParser = require("./javascript/JavascriptParser");
-const JavascriptParserApplyPlugin = require("./javascript/JavascriptParserApplyPlugin");
 
 const JsonModulesPlugin = require("./json/JsonModulesPlugin");
 const CircularModulesPlugin = require("./optimize/CircularModulesPlugin");
@@ -395,11 +392,6 @@ class WebpackOptionsApply extends OptionsApply {
 		}
 
 		new JavascriptModulesPlugin().apply(compiler);
-		new JavascriptParserApplyPlugin(
-			new ConstValueParserPlugin({
-				inlineExports: Boolean(options.optimization.inlineExports)
-			})
-		).apply(compiler);
 		new CircularModulesPlugin().apply(compiler);
 
 		new JsonModulesPlugin().apply(compiler);
@@ -648,11 +640,13 @@ class WebpackOptionsApply extends OptionsApply {
 
 			new InnerGraphPlugin().apply(compiler);
 		}
-		if (options.optimization.inlineExports) {
-			const InlineExportsPlugin = require("./optimize/InlineExportsPlugin");
 
-			new InlineExportsPlugin().apply(compiler);
-		}
+		const ConstExportsPlugin = require("./optimize/ConstExportsPlugin");
+
+		new ConstExportsPlugin({
+			inlineExports: options.optimization.inlineExports
+		}).apply(compiler);
+
 		if (options.optimization.mangleExports) {
 			const MangleExportsPlugin = require("./optimize/MangleExportsPlugin");
