+++ lib/dependencies/HarmonyExportDependencyParserPlugin.js
@@ -7,8 +7,8 @@
 
 const CompatibilityPlugin = require("../CompatibilityPlugin");
 const WebpackError = require("../errors/WebpackError");
-const { CONST_BINDING_TAG } = require("../javascript/ConstValueParserPlugin");
 const { getImportAttributes } = require("../javascript/JavascriptParser");
+const { CONST_BINDING_TAG } = require("../optimize/ConstExportsPlugin");
 const { getInnerGraphUtils } = require("../optimize/InnerGraph");
 const ConstDependency = require("./ConstDependency");
 const HarmonyExportExpressionDependency = require("./HarmonyExportExpressionDependency");
