+++ lib/optimize/ConstExportsPlugin.js
@@ -11,18 +11,35 @@ const {
 	JAVASCRIPT_MODULE_TYPE_ESM
 } = require("../ModuleTypeConstants");
 const BasicEvaluatedExpression = require("../javascript/BasicEvaluatedExpression");
-const { CONST_BINDING_TAG } = require("../javascript/ConstValueParserPlugin");
-const { InlinedUsedName, enableInlineExports } = require("./InlineExports");
+const { VariableInfoFlags } = require("../javascript/JavascriptParser");
+const {
+	InlinedUsedName,
+	enableInlineExports,
+	toInlinedValue
+} = require("./InlineExports");
 
 /** @typedef {import("./InlineExports").InlinedValue} InlinedValue */
 /** @typedef {import("../Compiler")} Compiler */
 /** @typedef {import("../ExportsInfo")} ExportsInfo */
 /** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
 /** @typedef {import("../javascript/JavascriptModule").JavascriptModuleBuildInfo} JavascriptModuleBuildInfo */
+/** @typedef {import("../Module").BuildInfo} BuildInfo */
 
-const PLUGIN_NAME = "InlineExportsPlugin";
+/** @typedef {{ inlineExports: boolean }} ConstValueParserPluginOptions */
 
-class InlineExportsPlugin {
+const PLUGIN_NAME = "ConstExportsPlugin";
+
+// Parser tag for top-level `const X = <primitive>` bindings
+const CONST_BINDING_TAG = Symbol("const binding");
+
+class ConstExportsPlugin {
+	/**
+	 * @param {ConstValueParserPluginOptions} options plugin options
+	 */
+	constructor(options) {
+		this.options = options;
+	}
+
 	/**
 	 * @param {Compiler} compiler the compiler instance
 	 * @returns {void}
@@ -31,14 +48,11 @@ class InlineExportsPlugin {
 		compiler.hooks.compilation.tap(
 			PLUGIN_NAME,
 			(compilation, { normalModuleFactory }) => {
-				// Lets HarmonyImportSpecifierDependency.getCondition skip the inline check
-				// (and stay unconditional) when this plugin is not active
-				enableInlineExports(compilation.moduleGraph);
 				/**
 				 * @param {JavascriptParser} parser the parser
 				 * @returns {void}
 				 */
-				const handler = (parser) => {
+				const handleInlineExports = (parser) => {
 					parser.hooks.program.tap(PLUGIN_NAME, () => {
 						const buildInfo =
 							/** @type {JavascriptModuleBuildInfo | undefined} */
@@ -72,12 +86,75 @@ class InlineExportsPlugin {
 							}
 						});
 				};
+
+				/**
+				 * @param {JavascriptParser} parser the parser
+				 * @returns {void}
+				 */
+				const handleConstValue = (parser) => {
+					// Only const is tracked; function/class names can be reassigned in sloppy mode.
+					// Re-exports always use getters: cross-module bindings may be mutable,
+					// and SideEffectsFlagPlugin can rewire connections skipping the template.
+					parser.hooks.preDeclarator.tap(
+						PLUGIN_NAME,
+						(declarator, statement) => {
+							// Detect top-level `const` declarations:
+							// - Tag ALL const bindings with CONST_BINDING_TAG (for const detection via tag system)
+							// - Carry inlined primitive value in tag data when eligible (for inline optimization)
+							if (statement.kind !== "const") return;
+							if (parser.scope.topLevelScope !== true) return;
+
+							if (declarator.id.type === "Identifier") {
+								let inlinedValue;
+								if (this.options.inlineExports && declarator.init) {
+									const evaluated = parser.evaluateExpression(declarator.init);
+									inlinedValue = toInlinedValue(evaluated);
+								}
+								parser.tagVariable(
+									declarator.id.name,
+									CONST_BINDING_TAG,
+									inlinedValue ? { value: inlinedValue } : {},
+									VariableInfoFlags.Normal
+								);
+							} else {
+								// Handle destructuring patterns (ObjectPattern, ArrayPattern, etc.)
+								parser.enterPattern(declarator.id, (name) => {
+									parser.tagVariable(
+										name,
+										CONST_BINDING_TAG,
+										{},
+										VariableInfoFlags.Normal
+									);
+								});
+							}
+						}
+					);
+				};
+
+				/**
+				 * @param {JavascriptParser} parser the parser
+				 * @returns {void}
+				 */
+				const handler = (parser) => {
+					handleConstValue(parser);
+					if (this.options.inlineExports) {
+						handleInlineExports(parser);
+					}
+				};
+
 				normalModuleFactory.hooks.parser
 					.for(JAVASCRIPT_MODULE_TYPE_AUTO)
 					.tap(PLUGIN_NAME, handler);
+
 				normalModuleFactory.hooks.parser
 					.for(JAVASCRIPT_MODULE_TYPE_ESM)
 					.tap(PLUGIN_NAME, handler);
+
+				if (!this.options.inlineExports) return;
+
+				// Lets HarmonyImportSpecifierDependency.getCondition skip the inline check
+				// (and stay unconditional) when this plugin is not active
+				enableInlineExports(compilation.moduleGraph);
 
 				const moduleGraph = compilation.moduleGraph;
 				compilation.hooks.optimizeDependencies.tap(PLUGIN_NAME, (modules) => {
@@ -129,4 +206,5 @@ class InlineExportsPlugin {
 	}
 }
 
-module.exports = InlineExportsPlugin;
+module.exports = ConstExportsPlugin;
+module.exports.CONST_BINDING_TAG = CONST_BINDING_TAG;
